Add files via upload
In this research paper, we are going to discuss about various supervised machine learning techniques used for the detection of breast cancer and evaluate them using various evaluation techniques. The aim is to discover which of the machine learning methods has the potential to detect the type/nature of tumour (benign or malignant) and do a comparative analysis/ evaluation of the various machine learning algorithms like Logistic Regression, SVM (support machine vector), DT (decision tree) and Random Forest with the Deep Learning.